+++ src/components/TeacherComponents/HeaderListing/HeaderListing.tsx
@@ -1,4 +1,5 @@
 import { FC } from "react";
+import { useFeatureFlagEnabled } from "posthog-js/react";
 import {
   OakFlex,
   OakHeading,
@@ -6,6 +7,9 @@ import {
   OakColorFilterToken,
   OakBox,
   OakInlineBanner,
+  OakP,
+  OakSecondaryButton,
+  OakTagFunctional,
 } from "@oaknational/oak-components";
 
 import UnitDownloadButton, {
@@ -48,10 +52,14 @@ export type HeaderListingProps = {
   programmeFactor: string;
   hasCurriculumDownload?: boolean;
   shareButton?: React.ReactNode;
+  copiedComponent?: React.ReactNode;
   unitDownloadFileId?: string;
   onUnitDownloadSuccess?: () => void;
   showRiskAssessmentBanner?: boolean;
+  isIncompleteUnit?: boolean;
   subjectDescriptionUnitListingData?: UnitListingData;
+  isUnitSaved?: boolean;
+  onSave?: () => void;
 };
 
 const HeaderListing: FC<HeaderListingProps> = (props) => {
@@ -70,10 +78,14 @@ const HeaderListing: FC<HeaderListingProps> = (props) => {
     tierTitle,
     yearTitle,
     shareButton,
+    copiedComponent,
     unitDownloadFileId,
     onUnitDownloadSuccess,
     showRiskAssessmentBanner,
+    isIncompleteUnit,
     subjectDescriptionUnitListingData,
+    isUnitSaved,
+    onSave,
   } = props;
 
   const isKeyStagesAvailable = keyStageSlug && keyStageTitle;
@@ -86,7 +98,11 @@ const HeaderListing: FC<HeaderListingProps> = (props) => {
     setDownloadError,
     setDownloadInProgress,
     downloadInProgress,
+    showIncompleteMessage,
+    setShowIncompleteMessage,
   } = useUnitDownloadButtonState();
+
+  const isSaveEnabled = useFeatureFlagEnabled("teacher-save-units");
 
   const bannersBlock = (
     <>
@@ -98,15 +114,33 @@ const HeaderListing: FC<HeaderListingProps> = (props) => {
             message="Sorry, download is not working. Please try again in a few minutes."
             icon="error"
           />
-        ) : showDownloadMessage ? (
-          <OakInlineBanner
-            isOpen={showDownloadMessage}
-            canDismiss
-            onDismiss={() => setShowDownloadMessage(false)}
-            type="neutral"
-            message="Downloads may take a few minutes on slower Wi-Fi connections."
-            $mb={"space-between-s"}
-          />
+        ) : showDownloadMessage || showIncompleteMessage ? (
+          <>
+            <OakInlineBanner
+              isOpen={showDownloadMessage}
+              canDismiss
+              onDismiss={() => setShowDownloadMessage(false)}
+              type="neutral"
+              message={
+                <OakP>
+                  Downloads may take a few minutes on slower Wi-Fi connections.
+                </OakP>
+              }
+              $mb={"space-between-s"}
+              $width="max-content"
+            />
+            {isIncompleteUnit && (
+              <OakInlineBanner
+                canDismiss
+                onDismiss={() => setShowIncompleteMessage(false)}
+                isOpen={showIncompleteMessage}
+                type="neutral"
+                message={<OakP>This unit is incomplete</OakP>}
+                $mb={"space-between-s"}
+                $width="max-content"
+              />
+            )}
+          </>
         ) : null}
       </OakBox>
       {showRiskAssessmentBanner && <RiskAssessmentBanner />}
@@ -163,20 +197,53 @@ const HeaderListing: FC<HeaderListingProps> = (props) => {
             <OakFlex $flexDirection="column" $gap="space-between-s">
               <OakFlex
                 $gap="space-between-s"
-                $flexDirection={["column", "row"]}
+                $height="max-content"
+                $flexWrap="wrap"
               >
                 {unitDownloadFileId && onUnitDownloadSuccess && (
                   <UnitDownloadButton
                     setDownloadError={setDownloadError}
                     setDownloadInProgress={setDownloadInProgress}
                     setShowDownloadMessage={setShowDownloadMessage}
+                    setShowIncompleteMessage={setShowIncompleteMessage}
                     downloadInProgress={downloadInProgress}
                     unitFileId={unitDownloadFileId}
                     onDownloadSuccess={onUnitDownloadSuccess}
+                    showNewTag={!isSaveEnabled}
                   />
                 )}
                 {shareButton}
+                {isSaveEnabled && onSave && (
+                  <OakSecondaryButton
+                    iconName={
+                      isUnitSaved ? "bookmark-filled" : "bookmark-outlined"
+                    }
+                    isTrailingIcon
+                    onClick={onSave}
+                    ph={["inner-padding-xs", "inner-padding-m"]}
+                    pv={["inner-padding-ssx", "inner-padding-ssx"]}
+                    $mb={[
+                      "space-between-none",
+                      "space-between-s",
+                      "space-between-none",
+                    ]}
+                    data-testid="save-unit-button"
+                  >
+                    <OakFlex $alignItems="center" $gap={"space-between-xs"}>
+                      <OakTagFunctional
+                        label="New"
+                        $background="mint"
+                        $color="text-primary"
+                        $font="heading-light-7"
+                        $pv={"inner-padding-none"}
+                        $display={["none", "inline"]}
+                      />
+                      {isUnitSaved ? "Saved" : "Save"}
+                    </OakFlex>
+                  </OakSecondaryButton>
+                )}
               </OakFlex>
+              {copiedComponent}
               <OakBox $display={["none", "block", "block"]}>
                 {bannersBlock}
               </OakBox>