+++ src/node-lib/curriculum-api-2023/queries/specialistLessonListing/specialistLessonListing.query.test.ts
@@ -30,6 +30,7 @@ jest.mock("../../sdk", () => {
             contains_copyright_content: false,
             unit_title: "Staying safe",
             order_in_unit: 2,
+            isUnpublished: false,
           },
           {
             unit_slug: "staying-safe-al-5556",
@@ -53,6 +54,7 @@ jest.mock("../../sdk", () => {
             contains_copyright_content: false,
             unit_title: "Staying safe",
             order_in_unit: 1,
+            isUnpublished: false,
           },
         ],
       }),