+++ src/utils/curriculum/isVisibleUnit.ts
@@ -1,25 +1,40 @@
-import { YearSelection } from "./types";
+import { CurriculumFilters } from "./types";
 
 import { CurriculumUnitsTabData } from "@/node-lib/curriculum-api-2023";
 
+export type PartialFilters = {
+  childSubjects?: CurriculumFilters["childSubjects"];
+  subjectCategories?: CurriculumFilters["subjectCategories"];
+  tiers?: CurriculumFilters["tiers"];
+  years: CurriculumFilters["years"];
+  threads: CurriculumFilters["threads"];
+};
+
 export function isVisibleUnit(
-  yearSelection: YearSelection,
+  filters: PartialFilters,
   year: string,
   unit: CurriculumUnitsTabData["units"][number],
 ) {
-  const s = yearSelection[year];
-  if (!s) {
+  if (!filters.years.includes(year)) {
     return false;
   }
   const filterBySubject =
-    !s.subject || s.subject.subject_slug === unit.subject_slug;
+    !filters.childSubjects?.[0] ||
+    filters.childSubjects[0] === unit.subject_slug;
+
   const filterBySubjectCategory =
-    s.subjectCategory?.id == -1 ||
+    !filters.subjectCategories?.[0] ||
+    (filters.subjectCategories.length > 0 &&
+      filters.subjectCategories?.[0] === "-1") ||
     unit.subjectcategories?.findIndex(
-      (subjectcategory) => subjectcategory.id === s.subjectCategory?.id,
+      (subjectcategory) =>
+        String(subjectcategory.id) === filters.subjectCategories?.[0],
     ) !== -1;
+
   const filterByTier =
-    !s.tier || !unit.tier_slug || s.tier?.tier_slug === unit.tier_slug;
+    !filters.tiers?.[0] ||
+    !unit.tier_slug ||
+    filters.tiers[0] === unit.tier_slug;
 
   return filterBySubject && filterBySubjectCategory && filterByTier;
 }